nze_hvac rubocop edits
